requirements.txt
flask
flask-cors
python-dotenv
joblib
pandas
numpy
supabase
langchain
huggingface-hub
ibm-watsonx-ai
langchain_community
langchain_ibm
sentence-transformers
openai
faiss-gpu
